refactor(nimbus): make mint_preview_experiment a factory fixture
Other tests in this suite (create_experiment in nimbus/conftest.py)
put experiment-setup wrappers behind a factory fixture, not a plain
module-level function. Doing the same for Fenix keeps the pattern
consistent.

This commit

* Moves mint_preview_experiment + wait_for_recipe from the test
  module to experimenter/tests/integration/nimbus/android/conftest.py,
  packaging them into a single factory fixture that:
  - Creates the draft via helpers.create_experiment
  - Transitions draft → preview via helpers.launch_to_preview
  - Polls /api/v6/experiments/{slug}/ until bucketConfig is populated
  - Returns the resolved recipe
* application_feature_ids is pulled from the existing parent conftest
  fixture rather than looked up directly.
* test_fenix_enrollment shrinks to the actual behavior it asserts:
  call the fixture to get a recipe, push via nimbus-cli, assert the
  log-state row shows up in logcat.
